collation: Introduce zh@collation=(unihan|pinyin|zhuyin|stroke)
They are ready to use by setting `$wgCategoryCollation` to one of the
valid options.

I am not sure about the output of the collation data generation script,
because it's never updated for new ICU versions, and a lot of characters
from the Supplementary Multilingual Plane would be added.
So, I just removed the Kanbun characters that can cause problems
manually, and which were missed by the initial script run.

Bug: T46667
Bug: T170049
Change-Id: I9eda1bd1acf2b4ffce5f351096a4bb96031956c1